+++ CHANGELOG.md
@@ -1,6 +1,5 @@
 # Change Log
 
 ## 0.1.3
-- Change Title/Description from context menu
-- Local CSS only for certain domains
-- Able to open/save Local CSS for root folder
+- Change Title/Description from context menu (except root folder)
+- Local CSS plugin support from context menu